Move select dialog related code to MultipleChoiceFieldView
