+++ gnd/src/main/java/com/google/android/gnd/ui/editobservation/SingleSelectDialogFactory.java
@@ -33,15 +33,15 @@ import java8.util.Optional;
 import java8.util.function.Consumer;
 
 // TODO: Replace with modal bottom sheet.
-class SingleSelectDialogFactory {
+public class SingleSelectDialogFactory {
 
   private Context context;
 
-  SingleSelectDialogFactory(Context context) {
+  public SingleSelectDialogFactory(Context context) {
     this.context = context;
   }
 
-  AlertDialog create(
+  public AlertDialog create(
       Field field,
       Optional<Response> initialValue,
       Consumer<Optional<Response>> valueChangeCallback) {
